# config/swaggerUser.json
{
    "swagger": "2.0",
    "info": {
        "title": "CSF Healthcare",
        "description": "CSF Healthcare api documentation for users",
        "version": "1.0"
    },
    "produces": [
        "application/json"
    ],
    "host": "192.168.102.76:3002",
    "basePath": "",
    "paths": {
...
        "/api/workingHours": {
            "put": {
                "x-swagger-router-controller": "middleware-name1",
                "operationId": "workingHours",
                "tags": [
                    "2 User APIs"
                ],
                "description": "working hours of professional/facility",
                "parameters": [
                    {
                        "name": "authorization",
                        "in": "header",
                        "description": "authorization is required",
                        "required": true,
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "language",
                        "in": "header",
                        "required": true,
                        "type": "string",
                        "enum": [
                            "en",
                            "ar"
                          ]
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "description": "working hours array like [{'Working':true,'startTime':'12:30 AM','endTime':'05:00 PM'}]",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "role": {
                                    "type": "string",
                                    "required": true,
                                    "description": "role is required [PROFESSIONAL / FACILITY]",
                                    "value": "PROFESSIONAL"
                                },
                                "workingHours": {
                                    "type": "array",
                                    "required": true,
                                    "collectionFormat": "multi",
                                    "items": {
                                        "type": "object"
                                    }
                                },
                                "isWholeWeekWorking": {
                                    "type": "boolean",
                                    "required": true
                                }
                            },
                            "example": {
                                "workingHours": [{
                                    "working":true,
                                    "startTime": "12:30 AM",
                                    "endTime": "05:00 PM"
                                }]
                            }
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/professionalFacilities": {
            "get": {
                "x-swagger-router-controller": "middleware-name1",
                "operationId": "professionalFacilities",
                "tags": [
                    "2 User APIs"
                ],
                "description": "professional facilities",
                "parameters": [
                    {
                        "name": "authorization",
                        "in": "header",
                        "description": "authorization is required",
                        "required": true,
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "language",
                        "in": "header",
                        "required": true,
                        "type": "string",
                        "enum": [
                            "en",
                            "ar"
                          ]
                    },
                    {
                        "name": "long",
                        "in": "query",
                        "required": true,
                        "description": "professional location longitude",
                        "type": "number",
                        "value": 0
                    },
                    {
                        "name": "lat",
                        "in": "query",
                        "required": true,
                        "description": "professional location latitude",
                        "type": "number",
                        "value": 0
                    },
                    {
                        "name": "userId",
                        "in": "query",
                        "required": true,
                        "description": "userId is required",
                        "type": "string",
                        "value": ""
                    }
                ],
                "responses": {}
            }
        },
...
    }
}